pyproject.toml
[project]
name = "translated-docs"
version = "0.1.0"
description = "Laravel 문서 번역 자동화 도구"
authors = [
    {name = "kimchanhyung98", email = "kimchanhyung98@users.noreply.github.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "openai",
    "requests",
    "tiktoken",
]

[project.scripts]
translate = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils", "src"]

[dependency-groups]
dev = []
